adapt "cast *id types to unsigned long when printing"
{u|g}id_t types vary by platform so cast to u_long and use %lu when printing them to prevent warnings.

+++ sftp-common.c
@@ -222,9 +222,9 @@ ls_file(const char *name, const struct stat *st, int remote, int si_units)
 
 	strmode(st->st_mode, mode);
 	if (remote) {
-		snprintf(ubuf, sizeof ubuf, "%u", (u_int)st->st_uid);
+		snprintf(ubuf, sizeof ubuf, "%lu", (u_long)st->st_uid);
 		user = ubuf;
-		snprintf(gbuf, sizeof gbuf, "%u", (u_int)st->st_gid);
+		snprintf(gbuf, sizeof gbuf, "%lu", (u_long)st->st_gid);
 		group = gbuf;
 		strlcpy(lc, "?", sizeof(lc));
 	} else {
